# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API Sandbox CESFAM",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/medications": {
      "get": {
        "tags": [
          "Medicamentos"
        ],
        "summary": "List Medications",
        "operationId": "list_medications_api_v1_medications_get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/medications/stock-summary": {
      "get": {
        "tags": [
          "Medicamentos"
        ],
        "summary": "Stock Summary",
        "operationId": "stock_summary_api_v1_medications_stock_summary_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/medications/low-stock": {
      "get": {
        "tags": [
          "Medicamentos"
        ],
        "summary": "Low Stock",
        "operationId": "low_stock_api_v1_medications_low_stock_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Medication"
                  },
                  "title": "Response Low Stock Api V1 Medications Low Stock Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/medications/{medication_id}": {
      "get": {
        "tags": [
          "Medicamentos"
        ],
        "summary": "Get Medication",
        "operationId": "get_medication_api_v1_medications__medication_id__get",
        "parameters": [
          {
            "name": "medication_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Medication Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedicationDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/medications/{medication_id}/batches": {
      "get": {
        "tags": [
          "Medicamentos"
        ],
        "summary": "List Batches",
        "operationId": "list_batches_api_v1_medications__medication_id__batches_get",
        "parameters": [
          {
            "name": "medication_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Medication Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Batch"
                  },
                  "title": "Response List Batches Api V1 Medications  Medication Id  Batches Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Medicamentos"
        ],
        "summary": "Add Batch",
        "operationId": "add_batch_api_v1_medications__medication_id__batches_post",
        "parameters": [
          {
            "name": "medication_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Medication Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Batch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Batch": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "medicationId": {
            "type": "string",
            "title": "Medicationid"
          },
          "batchNumber": {
            "type": "string",
            "title": "Batchnumber"
          },
          "arrivalDate": {
            "type": "string",
            "format": "date",
            "title": "Arrivaldate"
          },
          "expirationDate": {
            "type": "string",
            "format": "date",
            "title": "Expirationdate"
          },
          "initialQuantity": {
            "type": "integer",
            "title": "Initialquantity"
          },
          "availableQuantity": {
            "type": "integer",
            "title": "Availablequantity"
          }
        },
        "type": "object",
        "required": [
          "id",
          "medicationId",
          "batchNumber",
          "arrivalDate",
          "expirationDate",
          "initialQuantity",
          "availableQuantity"
        ],
        "title": "Batch"
      },
...
      "BatchCreate": {
        "properties": {
          "batchNumber": {
            "type": "string",
            "title": "Batchnumber"
          },
          "expirationDate": {
            "type": "string",
            "format": "date",
            "title": "Expirationdate"
          },
          "initialQuantity": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Initialquantity"
          }
        },
        "type": "object",
        "required": [
          "batchNumber",
          "expirationDate",
          "initialQuantity"
        ],
        "title": "BatchCreate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Medication": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "manufacturer": {
            "type": "string",
            "title": "Manufacturer"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "components": {
            "type": "string",
            "title": "Components"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "packaging": {
            "type": "string",
            "title": "Packaging"
          },
          "minStock": {
            "type": "integer",
            "title": "Minstock"
          },
          "stock": {
            "$ref": "#/components/schemas/Stock"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "description",
          "manufacturer",
          "type",
          "components",
          "content",
          "packaging",
          "minStock",
          "stock"
        ],
        "title": "Medication"
      },
      "MedicationDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "manufacturer": {
            "type": "string",
            "title": "Manufacturer"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "components": {
            "type": "string",
            "title": "Components"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "packaging": {
            "type": "string",
            "title": "Packaging"
          },
          "minStock": {
            "type": "integer",
            "title": "Minstock"
          },
          "stock": {
            "$ref": "#/components/schemas/Stock"
          },
          "batches": {
            "items": {
              "$ref": "#/components/schemas/Batch"
            },
            "type": "array",
            "title": "Batches"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "description",
          "manufacturer",
          "type",
          "components",
          "content",
          "packaging",
          "minStock",
          "stock",
          "batches"
        ],
        "title": "MedicationDetail"
      },
...
      "Stock": {
        "properties": {
          "availableQuantity": {
            "type": "integer",
            "title": "Availablequantity"
          },
          "reservedQuantity": {
            "type": "integer",
            "title": "Reservedquantity"
          },
          "physicalQuantity": {
            "type": "integer",
            "title": "Physicalquantity"
          }
        },
        "type": "object",
        "required": [
          "availableQuantity",
          "reservedQuantity",
          "physicalQuantity"
        ],
        "title": "Stock"
      },
      "StockSummary": {
        "properties": {
          "available": {
            "type": "integer",
            "title": "Available"
          },
          "lowStock": {
            "type": "integer",
            "title": "Lowstock"
          },
          "outOfStock": {
            "type": "integer",
            "title": "Outofstock"
          },
          "totalMedications": {
            "type": "integer",
            "title": "Totalmedications"
          }
        },
        "type": "object",
        "required": [
          "available",
          "lowStock",
          "outOfStock",
          "totalMedications"
        ],
        "title": "StockSummary"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}